pyproject.toml
[project]
name = "artidicia"
version = "0.1.0"
description = "Artidicia - The Artificial Serendipity Engine | Where AI hallucination becomes creative art"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.21",
    "jinja2>=3.1.6",
    "ollama>=0.6.1",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "streamlit>=1.51.0",
    "watchdog>=6.0.0",
]
